Use size_t rather than ssize_t for ec_mac_print.
Since global.h is included from usermode code, which doesn't know about ssize_t
without an additional header, and because there's no reason for this code to be
using ssize_t or off_t in the first place, let's just use size_t throughout.

+++ master/module.c
@@ -239,12 +239,12 @@ int ec_mac_equal(
  *
  * \return number of bytes written.
  */
-ssize_t ec_mac_print(
+size_t ec_mac_print(
         const uint8_t *mac, /**< MAC address */
         char *buffer /**< Target buffer. */
         )
 {
-    off_t off = 0;
+    size_t off = 0;
     unsigned int i;
 
     for (i = 0; i < ETH_ALEN; i++) {
